--- a/.rstcheck.cfg
+++ b/.rstcheck.cfg
@@ -1,0 +1,4 @@
+[rstcheck]
+report=info
+ignore_language=rst
+ignore_messages=(Duplicate (ex|im)plicit target.*|Hyperlink target ".*" is not referenced\.$)