--- a/CODEOWNERS
+++ b/CODEOWNERS
@@ -1,0 +1,53 @@
+# https://help.github.com/en/github/creating-cloning-and-archiving-repositories/about-code-owners
+
+# SECTION: Owner(s) for everything in the repo, unless a later match takes precedence
+# FILE PATTERN                              OWNER(S)
+*                                           @saltstack-formulas/wg
+
+# SECTION: Owner(s) for specific directories
+# FILE PATTERN                              OWNER(S)
+
+# SECTION: Owner(s) for files/directories related to `semantic-release`
+# FILE PATTERN                              OWNER(S)
+/.github/workflows/                         @saltstack-formulas/ssf
+/bin/install-hooks                          @saltstack-formulas/ssf
+/bin/kitchen                                @saltstack-formulas/ssf
+/docs/AUTHORS.rst                           @saltstack-formulas/ssf
+/docs/CHANGELOG.rst                         @saltstack-formulas/ssf
+/docs/TOFS_pattern.rst                      @saltstack-formulas/ssf
+/*/_mapdata/                                @saltstack-formulas/ssf
+/*/libmapstack.jinja                        @saltstack-formulas/ssf
+/*/libmatchers.jinja                        @saltstack-formulas/ssf
+/*/libsaltcli.jinja                         @saltstack-formulas/ssf
+/*/libtofs.jinja                            @saltstack-formulas/ssf
+/*/map.jinja                                @saltstack-formulas/ssf
+/test/integration/**/_mapdata.rb            @saltstack-formulas/ssf
+/test/integration/**/libraries/system.rb    @saltstack-formulas/ssf
+/test/integration/**/inspec.yml             @saltstack-formulas/ssf
+/test/integration/**/README.md              @saltstack-formulas/ssf
+/test/salt/pillar/top.sls                   @saltstack-formulas/ssf
+/.gitignore                                 @saltstack-formulas/ssf
+/.cirrus.yml                                @saltstack-formulas/ssf
+/.gitlab-ci.yml                             @saltstack-formulas/ssf
+/.pre-commit-config.yaml                    @saltstack-formulas/ssf
+/.rstcheck.cfg                              @saltstack-formulas/ssf
+/.rubocop.yml                               @saltstack-formulas/ssf
+/.salt-lint                                 @saltstack-formulas/ssf
+/.travis.yml                                @saltstack-formulas/ssf
+/.yamllint                                  @saltstack-formulas/ssf
+/AUTHORS.md                                 @saltstack-formulas/ssf
+/CHANGELOG.md                               @saltstack-formulas/ssf
+/CODEOWNERS                                 @saltstack-formulas/ssf
+/commitlint.config.js                       @saltstack-formulas/ssf
+/FORMULA                                    @saltstack-formulas/ssf
+/Gemfile                                    @saltstack-formulas/ssf
+/Gemfile.lock                               @saltstack-formulas/ssf
+/kitchen.yml                                @saltstack-formulas/ssf
+/kitchen.vagrant.yml                        @saltstack-formulas/ssf
+/kitchen.windows.yml                        @saltstack-formulas/ssf
+/pre-commit_semantic-release.sh             @saltstack-formulas/ssf
+/release-rules.js                           @saltstack-formulas/ssf
+/release.config.js                          @saltstack-formulas/ssf
+
+# SECTION: Owner(s) for specific files
+# FILE PATTERN                              OWNER(S)