--- a/commitlint.config.js
+++ b/commitlint.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    extends: ['@commitlint/config-conventional'],
+    rules: {
+        'body-max-line-length': [2, 'always', 120],
+        'footer-max-line-length': [2, 'always', 120],
+        'header-max-length': [2, 'always', 72],
+    },
+};