--- a/moosefs/cgi/cleaned.sls
+++ b/moosefs/cgi/cleaned.sls
@@ -1,0 +1,5 @@
+# -*- mode: salt; coding: utf-8 -*-
+# vim: ft=sls
+
+include:
+  - .package.cleaned