--- a/moosefs/cgi/init.sls
+++ b/moosefs/cgi/init.sls
@@ -1,0 +1,5 @@
+# -*- mode: salt; coding: utf-8 -*-
+# vim: ft=sls
+
+include:
+  - .package