--- a/moosefs/cgiserv/config/init.sls
+++ b/moosefs/cgiserv/config/init.sls
@@ -1,0 +1,5 @@
+# -*- mode: salt; coding: utf-8 -*-
+# vim: ft=sls
+
+include:
+  - .moosefs-cgiserv