--- a/moosefs/cgiserv/init.sls
+++ b/moosefs/cgiserv/init.sls
@@ -1,0 +1,7 @@
+# -*- mode: salt; coding: utf-8 -*-
+# vim: ft=sls
+
+include:
+  - .package
+  - .config
+  - .service