--- a/moosefs/cgiserv/package/init.sls
+++ b/moosefs/cgiserv/package/init.sls
@@ -1,0 +1,5 @@
+# -*- mode: salt; coding: utf-8 -*-
+# vim: ft=sls
+
+include:
+  - .installed