--- a/moosefs/chunkserver/cleaned.sls
+++ b/moosefs/chunkserver/cleaned.sls
@@ -1,0 +1,7 @@
+# -*- mode: salt; coding: utf-8 -*-
+# vim: ft=sls
+
+include:
+  - .service.cleaned
+  - .config.cleaned
+  - .package.cleaned