--- a/moosefs/chunkserver/config/files/default/mfshdd.cfg.jinja
+++ b/moosefs/chunkserver/config/files/default/mfshdd.cfg.jinja
@@ -1,0 +1,51 @@
+# File managed by SaltStack at <{{ source }}>
+# Your changes will be overwritten.
+#
+# mfshdd.cfg(5)
+#
+# This file keeps definitions of mounting points (paths) of hard drives to use with chunk server.
+# A path may begin with extra characters which swiches additional options:
+#  - '*' means that this hard drive is 'marked for removal' and all data will be replicated to other hard drives (usually on other chunkservers)
+#  - '<' means that all data from this hard drive should be moved to other local hard drives
+#  - '>' means that all data from other local hard drives should be moved to this hard drive
+#  - '~' means that significant change of total blocks count will not mark this drive as damaged
+# If there are both '<' and '>' drives then data will be moved only between these drives
+# It is possible to specify optional space limit (after each mounting point), there are two ways of doing that:
+#  - set space to be left unused on a hard drive (this overrides the default setting from mfschunkserver.cfg)
+#  - limit space to be used on a hard drive
+# Space limit definition: [0-9]*(.[0-9]*)?([kMGTPE]|[KMGTPE]i)?B?, add minus in front for the first option.
+#
+# Examples:
+#
+# use hard drive '/mnt/hd1' with default options:
+#/mnt/hd1
+#
+# use hard drive '/mnt/hd2', but replicate all data from it:
+#*/mnt/hd2
+#
+# use hard drive '/mnt/hd3', but try to leave 5GiB on it:
+#/mnt/hd3 -5GiB
+#
+# use hard drive '/mnt/hd4', but use only 1.5TiB on it:
+#/mnt/hd4 1.5TiB
+#
+# use hard drive '/mnt/hd5', but fill it up using data from other drives
+#>/mnt/hd5
+#
+# use hard drive '/mnt/hd6', but move all data to other hard drives
+#</mnt/hd6
+#
+# use hard drive '/mnt/hd7', but ignore significant change of hard drive total size (e.g. compressed file systems)
+#~/mnt/hd7
+#
+# mount points of HDD drives
+#
+{%- for hdd, param in chunkserver | traverse("config:hdds", {}) | dictsort %}
+{{ "{}{} {}".format(
+  param.get("option") | default("", boolean=True),
+  hdd,
+  param.get("size_limit") | default("", boolean=True)
+  )
+  | trim
+}}
+{%- endfor %}