--- a/moosefs/files/metadata.mfs
+++ b/moosefs/files/metadata.mfs
@@ -1,1 +1,0 @@
-MFSM NEW