… .gitlab-ci.yml
… .pre-commit-config.yaml
… .rubocop.yml
… Gemfile.lock
… README.rst
… bin/kitchen
… docs/README.rst
… docs/TOFS_pattern.rst
… docs/map.jinja.rst
… kitchen.yml
… moosefs/chunk.sls
… moosefs/chunkserver/config/init.sls
… moosefs/chunkserver/config/mfschunkserver.sls
… moosefs/chunkserver/config/mfshdd.sls
… moosefs/chunkserver/init.sls
… moosefs/chunkserver/package/cleaned.sls
… moosefs/chunkserver/package/init.sls
… moosefs/chunkserver/package/installed.sls
… moosefs/chunkserver/service/cleaned.sls
… moosefs/chunkserver/service/init.sls
… moosefs/chunkserver/service/running.sls
… moosefs/cleaned.sls
… moosefs/cli/cleaned.sls
… moosefs/cli/init.sls
… moosefs/cli/package/cleaned.sls
… moosefs/cli/package/init.sls
… moosefs/cli/package/installed.sls
… moosefs/client.sls
… moosefs/client/cleaned.sls
… moosefs/client/config/cleaned.sls