<!--
Please fill in this PR template to make it easier to review and merge.

It has been designed so that a lot of it can be completed *after* submitting it,
e.g. filling in the checklists.

Notes:
1. Please keep the PR as small as is practicable; the larger a PR gets, the harder it becomes to review and the more time it requires to get it merged.
2. Similarly, please avoid PRs that cover more than one type; it should be a _bug fix_ *OR* a _new feature_ *OR* a _refactor_, etc.
3. Please direct questions to the [`#formulas` channel on Slack](https://saltstackcommunity.slack.com/messages/C7LG8SV54/), which is bridged to `#saltstack-formulas` on Freenode.
4. Feel free to suggest improvements to this template by reporting an issue or submitting a PR.  The source of this template is:
  - https://github.com/saltstack-formulas/.github/blob/master/.github/pull_request_template.md
-->

### PR progress checklist (to be filled in by reviewers)
<!-- Please leave this checklist for reviewers to tick as they work through the PR. -->

- [ ] Changes to documentation are appropriate (or tick if not required)
- [ ] Changes to tests are appropriate (or tick if not required)
- [ ] Reviews completed

---

### What type of PR is this?
<!-- Please tick each box that is relevant (after creating the PR). -->

#### Primary type
<!-- There really should be only *one* of these types ticked for each PR. -->

- [ ] `[build]`    Changes related to the build system
- [ ] `[chore]`    Changes to the build process or auxiliary tools and libraries such as documentation generation
- [ ] `[ci]`       Changes to the continuous integration configuration
- [ ] `[feat]`     A new feature
- [ ] `[fix]`      A bug fix
- [ ] `[perf]`     A code change that improves performance
- [ ] `[refactor]` A code change that neither fixes a bug nor adds a feature
- [ ] `[revert]`   A change used to revert a previous commit
- [ ] `[style]`    Changes that do not affect the meaning of the code (white-space, formatting, missing semi-colons, etc.)

#### Secondary type
<!-- Most PRs should include all of the following types as well. -->

- [ ] `[docs]`     Documentation changes
- [ ] `[test]`     Adding missing or correcting existing tests

### Does this PR introduce a `BREAKING CHANGE`?
<!-- If so, change the following to a `Yes` and explain what the breaking changes are. -->
<!-- If there are multiple breaking changes, list them all. -->

YES

- init: you must set master:enabled, metalogger:enabled, chunkserver:enabled, cgiserv:enabled, client:enabled, cli:enabled and netdump:enabled to install each component with the top level state
- cgiserv: the cgiserv was provided by moosefs.master and is now split in moosefs.cgiserv.package, moosefs.cgiserv.config and moosefs.cgiserv.service
- client: moosefs.client is split in moosefs.client.package and moosefs.client.config
- chunkserver: moosefs.chunkserver is split in moosefs.chunkserver.package, moosefs.chunkserver.config and moosefs.chunkserver.service
- metalogger: moosefs.metalogger is split in moosefs.metalogger.package, moosefs.metalogger.config and moosefs.metalogger.service
- master: moosefs.master is split in moosefs.master.package, moosefs.master.config and moosefs.master.service
- map: map.jinja now exports a generic mapdata variable
- map: The per grain parameter values are now under moosefs/parameters/

### Related issues and/or pull requests
<!-- Please link any related issues/PRs here, especially any issues that are closed by this PR. -->

saltstack-formulas/template-formula#21

### Describe the changes you're proposing
<!-- A clear and concise description of what you have implemented. -->
<!-- Consider explaining each commit if they cover different aspects of the proposed changes. -->

I update the formula to new community standards:

- update to v5 `map.jinja`
- split each component into subdirectory and provide a `.cleaned` to remove them
- use packages from distribution, only support [Debian 11, Ubuntu 22.04 and FreeBSD 12/13](https://pkgs.org/search/?q=moosefs)
- tested for Debian 11 and Ubuntu 22.04
- use [proposed `libs/` move](saltstack-formulas/template-formula#252)

### Pillar / config required to test the proposed changes
<!-- Provide links to the SLS files and/or relevant configs (be sure to remove sensitive info). -->

- [Kitchen test pillars for all in one setup](test/salt/pillar/default.sls)

### Debug log showing how the proposed changes work
<!-- Include a debug log showing how these changes work, e.g. using `salt-minion -l debug`. -->
<!-- Alternatively, linking to Kitchen debug logs is useful, e.g. via. Travis CI. -->
<!-- Most useful is providing a passing InSpec test, which can be used to verify any proposed changes. -->

<details><summary><code>./bin/kitchen converge default-debian-11-3004-1-py3</code></summary>

```
-----> Starting Test Kitchen (v3.3.1)
-----> Converging <default-debian-11-3004-1-py3>...
       Preparing files for transfer
       Preparing salt-minion
       Preparing pillars into /srv/pillar
       Copying pillars from test/salt/pillar to /srv/pillar
       Preparing formula: moosefs from /root/moosefs-formula
       Preparing state_top
       Preparing scripts into /etc/salt/scripts
       You asked for latest and you have 3004.1 installed, sweet!
       Transferring files to <default-debian-11-3004-1-py3>
       Install External Dependencies
       Content of /tmp/kitchen//srv/salt :
       total 16
       drwxr-xr-x  3 kitchen kitchen 4096 Jan  4 10:31 .
       drwxr-xr-x  4 kitchen kitchen 4096 Jan  4 10:31 ..
       drwxr-xr-x 14 kitchen kitchen 4096 Jan  4 10:31 moosefs
       -rw-r--r--  1 kitchen kitchen   50 Jan  4 10:55 top.sls
       [DEBUG   ] Reading configuration from /tmp/kitchen/etc/salt/minion
       [DEBUG   ] Using cached minion ID from /tmp/kitchen/etc/salt/minion_id: moosefs.example.net
       [DEBUG   ] Using importlib_metadata to load entry points
       [DEBUG   ] Override  __grains__: <module 'salt.loaded.int.log_handlers.sentry_mod' from '/usr/lib/python3/dist-packages/salt/log/handlers/sentry_mod.py'>
       [DEBUG   ] Configuration file path: /tmp/kitchen/etc/salt/minion
       [WARNING ] Insecure logging configuration detected! Sensitive data may be logged.
       [DEBUG   ] Grains refresh requested. Refreshing grains.
       [DEBUG   ] Reading configuration from /tmp/kitchen/etc/salt/minion
       [DEBUG   ] Override  __utils__: <module 'salt.loaded.int.grains.zfs' from '/usr/lib/python3/dist-packages/salt/grains/zfs.py'>
       [DEBUG   ] Elapsed time getting FQDNs: 0.04139208793640137 seconds
       [DEBUG   ] LazyLoaded zfs.is_supported
       [DEBUG   ] Determining pillar cache
       [DEBUG   ] Using importlib_metadata to load entry points
       [DEBUG   ] LazyLoaded jinja.render
       [DEBUG   ] LazyLoaded yaml.render
       [DEBUG   ] compile template: /tmp/kitchen/srv/pillar/top.sls
       [DEBUG   ] Jinja search path: ['/tmp/kitchen/srv/pillar', '/tmp/kitchen/srv/spm/pillar']
       [PROFILE ] Time (in seconds) to render '/tmp/kitchen/srv/pillar/top.sls' using 'jinja' renderer: 0.009363174438476562
       [DEBUG   ] Rendered data from file: /tmp/kitchen/srv/pillar/top.sls:
       # -*- coding: utf-8 -*-
       # vim: ft=yaml
       ---
       base:
         '*':
           - default
       ...
       […]
       [DEBUG   ] Results of YAML rendering: 
       OrderedDict([('moosefs-cli-package-installed-pkg-installed', OrderedDict([('pkg.installed', [OrderedDict([('name', 'moosefs-cli')])])]))])
       [PROFILE ] Time (in seconds) to render '/tmp/kitchen/var/cache/salt/minion/files/base/moosefs/cli/package/installed.sls' using 'yaml' renderer: 0.00042629241943359375
       [DEBUG   ] LazyLoaded file.managed
       [INFO    ] Running state [/tmp/salt_mapdata_dump.yaml] at time 10:55:51.385699
       [INFO    ] Executing state file.managed for [/tmp/salt_mapdata_dump.yaml]
       [DEBUG   ] LazyLoaded file.source_list
       [DEBUG   ] LazyLoaded cp.hash_file
       [DEBUG   ] Using importlib_metadata to load entry points
       [DEBUG   ] LazyLoaded roots.envs
       [DEBUG   ] Could not LazyLoad roots.init: 'roots.init' is not available.
       [DEBUG   ] In saltenv 'base', looking at rel_path 'moosefs/_mapdata/_mapdata.jinja' to resolve 'salt://moosefs/_mapdata/_mapdata.jinja'
       [DEBUG   ] In saltenv 'base', ** considering ** path '/tmp/kitchen/var/cache/salt/minion/files/base/moosefs/_mapdata/_mapdata.jinja' to resolve 'salt://moosefs/_mapdata/_mapdata.jinja'
       [DEBUG   ] Jinja search path: ['/tmp/kitchen/var/cache/salt/minion/files/base']
       [INFO    ] File /tmp/salt_mapdata_dump.yaml is in the correct state
       [INFO    ] Completed state [/tmp/salt_mapdata_dump.yaml] at time 10:55:51.454203 (duration_in_ms=68.505)
       [DEBUG   ] LazyLoaded pkg.install
       [DEBUG   ] LazyLoaded pkg.installed
       [DEBUG   ] LazyLoaded path.which
       [DEBUG   ] Override  __salt__: <module 'salt.loaded.int.module.dracr' from '/usr/lib/python3/dist-packages/salt/modules/dracr.py'>
       [DEBUG   ] Reading configuration from /etc/salt/minion
       [DEBUG   ] Guessing ID. The id can be explicitly set in /etc/salt/minion
       [DEBUG   ] Found minion id from generate_minion_id(): moosefs.example.net
       [DEBUG   ] Using importlib_metadata to load entry points
       [DEBUG   ] Could not LazyLoad idem.hub: 'idem' __virtual__ returned False: No module named 'pop'
       [DEBUG   ] Override  __salt__: <module 'salt.loaded.int.module.kubeadm' from '/usr/lib/python3/dist-packages/salt/modules/kubeadm.py'>
       [DEBUG   ] Override  __salt__: <module 'salt.loaded.int.module.udev' from '/usr/lib/python3/dist-packages/salt/modules/udev.py'>
       [DEBUG   ] Override  __pillar__: <module 'salt.loaded.int.module.virtualenv_mod' from '/usr/lib/python3/dist-packages/salt/modules/virtualenv_mod.py'>
       [DEBUG   ] DSC: Only available on Windows systems
       [DEBUG   ] Module PSGet: Only available on Windows systems
       [DEBUG   ] Could not LazyLoad pkg.ex_mod_init: 'pkg.ex_mod_init' is not available.
       [INFO    ] Running state [moosefs-master] at time 10:55:52.996160
       [INFO    ] Executing state pkg.installed for [moosefs-master]
       [DEBUG   ] Could not LazyLoad pkg.resolve_capabilities: 'pkg.resolve_capabilities' is not available.
       [INFO    ] Executing command dpkg-query in directory '/home/kitchen'
       [INFO    ] All specified packages are already installed
       [INFO    ] Completed state [moosefs-master] at time 10:55:53.032869 (duration_in_ms=36.708)
       [INFO    ] Running state [/etc/mfs/mfsmaster.cfg] at time 10:55:53.033745
       [INFO    ] Executing state file.managed for [/etc/mfs/mfsmaster.cfg]
       [DEBUG   ] Returning file list from cache: age=8 cache_time=20 /tmp/kitchen/var/cache/salt/minion/file_lists/roots/base.p
       [DEBUG   ] Returning file list from cache: age=8 cache_time=20 /tmp/kitchen/var/cache/salt/minion/file_lists/roots/base.p
       [DEBUG   ] In saltenv 'base', looking at rel_path 'moosefs/master/config/files/default/mfsmaster.cfg.jinja' to resolve 'salt://moosefs/master/config/files/default/mfsmaster.cfg.jinja'
       [DEBUG   ] In saltenv 'base', ** considering ** path '/tmp/kitchen/var/cache/salt/minion/files/base/moosefs/master/config/files/default/mfsmaster.cfg.jinja' to resolve 'salt://moosefs/master/config/files/default/mfsmaster.cfg.jinja'
       [DEBUG   ] Jinja search path: ['/tmp/kitchen/var/cache/salt/minion/files/base']
       [DEBUG   ] In saltenv 'base', looking at rel_path 'moosefs/libs/libconfig.jinja' to resolve 'salt://moosefs/libs/libconfig.jinja'
       [DEBUG   ] In saltenv 'base', ** considering ** path '/tmp/kitchen/var/cache/salt/minion/files/base/moosefs/libs/libconfig.jinja' to resolve 'salt://moosefs/libs/libconfig.jinja'
       [INFO    ] File /etc/mfs/mfsmaster.cfg is in the correct state
       [INFO    ] Completed state [/etc/mfs/mfsmaster.cfg] at time 10:55:53.091096 (duration_in_ms=57.35)
       [INFO    ] Running state [/etc/mfs/mfstopology.cfg] at time 10:55:53.091725
       [INFO    ] Executing state file.managed for [/etc/mfs/mfstopology.cfg]
       [DEBUG   ] Returning file list from cache: age=8 cache_time=20 /tmp/kitchen/var/cache/salt/minion/file_lists/roots/base.p
       [DEBUG   ] Returning file list from cache: age=8 cache_time=20 /tmp/kitchen/var/cache/salt/minion/file_lists/roots/base.p
       [DEBUG   ] In saltenv 'base', looking at rel_path 'moosefs/master/config/files/default/mfstopology.cfg.jinja' to resolve 'salt://moosefs/master/config/files/default/mfstopology.cfg.jinja'
       [DEBUG   ] In saltenv 'base', ** considering ** path '/tmp/kitchen/var/cache/salt/minion/files/base/moosefs/master/config/files/default/mfstopology.cfg.jinja' to resolve 'salt://moosefs/master/config/files/default/mfstopology.cfg.jinja'
       [DEBUG   ] Jinja search path: ['/tmp/kitchen/var/cache/salt/minion/files/base']
       [INFO    ] File /etc/mfs/mfstopology.cfg is in the correct state
       [INFO    ] Completed state [/etc/mfs/mfstopology.cfg] at time 10:55:53.103590 (duration_in_ms=11.865)
       [INFO    ] Running state [/etc/mfs/mfsexports.cfg] at time 10:55:53.104027
       [INFO    ] Executing state file.managed for [/etc/mfs/mfsexports.cfg]
       [DEBUG   ] Returning file list from cache: age=8 cache_time=20 /tmp/kitchen/var/cache/salt/minion/file_lists/roots/base.p
       [DEBUG   ] Returning file list from cache: age=8 cache_time=20 /tmp/kitchen/var/cache/salt/minion/file_lists/roots/base.p
       [DEBUG   ] In saltenv 'base', looking at rel_path 'moosefs/master/config/files/default/mfsexports.cfg.jinja' to resolve 'salt://moosefs/master/config/files/default/mfsexports.cfg.jinja'
       [DEBUG   ] In saltenv 'base', ** considering ** path '/tmp/kitchen/var/cache/salt/minion/files/base/moosefs/master/config/files/default/mfsexports.cfg.jinja' to resolve 'salt://moosefs/master/config/files/default/mfsexports.cfg.jinja'
       [DEBUG   ] Jinja search path: ['/tmp/kitchen/var/cache/salt/minion/files/base']
       [DEBUG   ] In saltenv 'base', looking at rel_path 'moosefs/libs/libconfig.jinja' to resolve 'salt://moosefs/libs/libconfig.jinja'
       [DEBUG   ] In saltenv 'base', ** considering ** path '/tmp/kitchen/var/cache/salt/minion/files/base/moosefs/libs/libconfig.jinja' to resolve 'salt://moosefs/libs/libconfig.jinja'
       [INFO    ] File /etc/mfs/mfsexports.cfg is in the correct state
       [INFO    ] Completed state [/etc/mfs/mfsexports.cfg] at time 10:55:53.128745 (duration_in_ms=24.717)
       [INFO    ] Running state [/etc/default/moosefs-master] at time 10:55:53.129211
       [INFO    ] Executing state file.managed for [/etc/default/moosefs-master]
       [DEBUG   ] Returning file list from cache: age=8 cache_time=20 /tmp/kitchen/var/cache/salt/minion/file_lists/roots/base.p
       [DEBUG   ] Returning file list from cache: age=8 cache_time=20 /tmp/kitchen/var/cache/salt/minion/file_lists/roots/base.p
       [DEBUG   ] In saltenv 'base', looking at rel_path 'moosefs/master/config/files/default/moosefs-master-default.jinja' to resolve 'salt://moosefs/master/config/files/default/moosefs-master-default.jinja'
       [DEBUG   ] In saltenv 'base', ** considering ** path '/tmp/kitchen/var/cache/salt/minion/files/base/moosefs/master/config/files/default/moosefs-master-default.jinja' to resolve 'salt://moosefs/master/config/files/default/moosefs-master-default.jinja'
       [DEBUG   ] Jinja search path: ['/tmp/kitchen/var/cache/salt/minion/files/base']
       [INFO    ] File /etc/default/moosefs-master is in the correct state
       [INFO    ] Completed state [/etc/default/moosefs-master] at time 10:55:53.142208 (duration_in_ms=12.997)
       [INFO    ] Running state [/var/lib/mfs/metadata.mfs] at time 10:55:53.142572
       [INFO    ] Executing state file.copy for [/var/lib/mfs/metadata.mfs]
       [DEBUG   ] LazyLoaded lxd.version
       [DEBUG   ] LazyLoaded acme.cert
       [DEBUG   ] LazyLoaded apache.config
       [DEBUG   ] LazyLoaded apache.a2enconf
       [DEBUG   ] LazyLoaded apache.a2enmod
       [DEBUG   ] LazyLoaded apache.a2ensite
       [DEBUG   ] LazyLoaded at.at
       [DEBUG   ] Could not LazyLoad augeas.execute: 'augeas.execute' is not available.
       [DEBUG   ] LazyLoaded azurearm_compute.availability_set_create_or_update
       [DEBUG   ] LazyLoaded azurearm_dns.zones_list_by_resource_group
       [DEBUG   ] LazyLoaded azurearm_network.check_ip_address_availability
       [DEBUG   ] LazyLoaded azurearm_resource.resource_group_check_existence
       [DEBUG   ] LazyLoaded boto3_elasticache.cache_cluster_exists
       [DEBUG   ] LazyLoaded boto3_elasticsearch.describe_elasticsearch_domain
       [DEBUG   ] LazyLoaded boto3_route53.find_hosted_zone
       [DEBUG   ] LazyLoaded boto3_sns.topic_exists
       [DEBUG   ] LazyLoaded boto_apigateway.describe_apis
       [DEBUG   ] LazyLoaded boto_asg.exists
       [DEBUG   ] LazyLoaded boto_cfn.exists
       [DEBUG   ] LazyLoaded boto_cloudfront.get_distribution
       [DEBUG   ] LazyLoaded boto_cloudtrail.exists
       [DEBUG   ] LazyLoaded boto_cloudwatch.get_alarm
       [DEBUG   ] LazyLoaded boto_cloudwatch_event.exists
       [DEBUG   ] LazyLoaded boto_cognitoidentity.describe_identity_pools
       [DEBUG   ] LazyLoaded boto_datapipeline.create_pipeline
       [DEBUG   ] LazyLoaded boto_dynamodb.exists
       [DEBUG   ] LazyLoaded boto_ec2.get_key
       [DEBUG   ] LazyLoaded boto_elasticache.exists
       [DEBUG   ] LazyLoaded boto_elasticsearch_domain.exists
       [DEBUG   ] LazyLoaded boto_elb.exists
       [DEBUG   ] LazyLoaded boto_elbv2.target_group_exists
       [DEBUG   ] LazyLoaded boto_iam.get_user
       [DEBUG   ] LazyLoaded boto_iam.role_exists
       [DEBUG   ] LazyLoaded boto_iot.policy_exists
       [DEBUG   ] LazyLoaded boto_kinesis.exists
       [DEBUG   ] LazyLoaded boto_kms.describe_key
       [DEBUG   ] LazyLoaded boto_lambda.function_exists
       [DEBUG   ] LazyLoaded boto_asg.exists
       [DEBUG   ] LazyLoaded boto_rds.exists
       [DEBUG   ] LazyLoaded boto_route53.get_record
       [DEBUG   ] LazyLoaded boto_s3.get_object_metadata
       [DEBUG   ] LazyLoaded boto_s3_bucket.exists
       [DEBUG   ] LazyLoaded boto_secgroup.exists
       [DEBUG   ] LazyLoaded boto_sns.exists
       [DEBUG   ] LazyLoaded boto_sqs.exists
       [DEBUG   ] LazyLoaded boto_vpc.exists
       [DEBUG   ] LazyLoaded bower.list
       [DEBUG   ] LazyLoaded chef.client
       [DEBUG   ] LazyLoaded chocolatey.install
       [DEBUG   ] LazyLoaded cimc.get_system_info
       [DEBUG   ] LazyLoaded cisconso.set_data_value
       [DEBUG   ] LazyLoaded cron.list_tab
       [DEBUG   ] LazyLoaded csf.exists
       [DEBUG   ] LazyLoaded cyg.list
       [DEBUG   ] LazyLoaded ddns.update
       [DEBUG   ] Could not LazyLoad debconf.show: 'debconf.show' is not available.
       [DEBUG   ] LazyLoaded chassis.cmd
       [DEBUG   ] Could not LazyLoad docker.version: 'docker.version' is not available.
       [DEBUG   ] Could not LazyLoad docker.version: 'docker.version' is not available.
       [DEBUG   ] Could not LazyLoad docker.version: 'docker.version' is not available.
       [DEBUG   ] Could not LazyLoad docker.version: 'docker.version' is not available.
       [DEBUG   ] LazyLoaded eselect.exec_action
       [DEBUG   ] LazyLoaded esxi.cmd
       [INFO    ] Executing command git in directory '/home/kitchen'
       [DEBUG   ] stdout: git version 2.30.2
       [DEBUG   ] LazyLoaded github.list_users
       [DEBUG   ] LazyLoaded glanceng.image_get
       [DEBUG   ] LazyLoaded glusterfs.list_volumes
       [DEBUG   ] LazyLoaded elasticsearch.exists
…